requirements.txt
absl-py==1.4.0
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
cftime==1.6.2
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
fqdn==1.5.1
future==0.18.3
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
interpolate==0.5.2
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jax==0.4.13
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.4
jsonschema==4.17.3
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
mistune==3.0.1
ml-dtypes==0.2.0
mpmath==1.3.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
psycopg2-binary==2.9.6
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
overrides==7.3.1
packaging==21.3
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
Pint==0.22
platformdirs==3.8.0
pooch==1.7.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.1.0
pyproj==3.6.0
pyrsistent==0.19.3
pyshp==2.3.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-telegram-bot==20.4
pytz==2023.3
pywin32==306
pywinpty==2.0.10
pyxlsb==1.0.10
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scipy==1.11.1
Send2Trash==1.8.2
shapely==2.0.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
spark==0.2.1
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
termcolor==2.3.0
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
virtualenv==20.23.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
wget==3.2
widgetsnbextension==4.0.7
wrapt==1.14.1
xarray==2023.6.0
xlrd==2.0.1
